requirements.txt
# PDF processing
pdfplumber==0.11.4
pdf2image==1.17.0
Pillow==10.4.0

# OCR
pytesseract==0.3.13

# LLM
anthropic==0.40.0

# Validation
pydantic==2.10.3

# Environment
python-dotenv==1.0.1

# Logging
loguru==0.7.3

# CLI
click==8.1.8
